+++ packages/affiliates/libs/affiliates/devmining.js
@@ -4,7 +4,7 @@ const Promise = require("bluebird");
 const assert = require("assert");
 const lodash = require("lodash");
 
-const { getWeb3 } = require("@uma/common");
+const { getWeb3, ConvertDecimals } = require("@uma/common");
 const web3 = getWeb3();
 const { toWei, toBN, fromWei } = web3.utils;
 
@@ -12,16 +12,17 @@ const { EmpBalancesHistory } = require("../processors");
 const { Prices } = require("../models");
 const { DecodeLog } = require("../contracts");
 
-module.exports = ({ queries, empAbi, coingecko, synthPrices, firstEmpDate }) => {
+module.exports = ({ queries, coingecko, synthPrices, firstEmpDate }) => {
   assert(queries, "requires queries class");
-  assert(empAbi, "requires empAbi");
   assert(coingecko, "requires coingecko api");
   assert(synthPrices, "requires synthPrices api");
 
   // use firstEmpDate as a history cutoff when querying for events. We can safely say no emps were deployed before Jan of 2020.
   firstEmpDate = firstEmpDate || moment("2020-01-01", "YYYY-MM-DD").valueOf();
 
-  async function getBalanceHistory(empAddress, start, end) {
+  async function getBalanceHistory(empAddress, start, end, empAbi) {
+    assert(empAddress, "requires empAddress");
+    assert(empAbi, "requires empAbi");
     // stream is a bit more optimal than waiting for entire query to return as array
     // We need all logs from beginning of time. This could be optimized by deducing or supplying
     // the specific emp start time to narrow down the query.
@@ -45,11 +46,15 @@ module.exports = ({ queries, empAbi, coingecko, synthPrices, firstEmpDate }) => 
 
     return balancesHistory;
   }
+  // This now requires an array of tuples in order to handle varying versions of the EMP
+  // [
+  //   [empAddress, empAbi]
+  // ]
   async function getAllBalanceHistories(empAddresses = [], start, end) {
     return Promise.reduce(
       empAddresses,
-      async (result, empAddress) => {
-        result.push([empAddress, await getBalanceHistory(empAddress, start, end)]);
+      async (result, [empAddress, empAbi]) => {
+        result.push([empAddress, await getBalanceHistory(empAddress, start, end, empAbi)]);
         return result;
       },
       []
@@ -127,10 +132,10 @@ module.exports = ({ queries, empAbi, coingecko, synthPrices, firstEmpDate }) => 
   function calculateValueFromUsd(...args) {
     // have to do this because eslint wont pass and wont ignore unused vars. These params need to be
     // identical to calculateValue so the calls can be interchangeable.
-    const [tokens, , syntheticPrice] = args;
-    return toBN(tokens.toString())
-      .mul(toBN(toWei(syntheticPrice.toFixed(18))))
-      .div(toBN(toWei("1")));
+    const [tokens, , syntheticPrice, , syntheticTokenDecimal] = args;
+    // EMP API has changed, now synthetics are not always 18 decimals. We have to normalize them to 18 before value calc.
+    const tokensNormalized = ConvertDecimals(syntheticTokenDecimal, 18, web3)(tokens.toString());
+    return tokensNormalized.mul(toBN(toWei(syntheticPrice.toFixed(18)))).div(toBN(toWei("1")));
   }
 
   function toChecksumAddress(address) {
@@ -141,11 +146,11 @@ module.exports = ({ queries, empAbi, coingecko, synthPrices, firstEmpDate }) => 
   function validateEmpInput(empValue) {
     assert(
       lodash.isArray(empValue),
-      "Each EMP whitelisted is expected to be an array in the form [empAddress, rewardAddress]"
+      "Each EMP whitelisted is expected to be an array in the form [empAddress, rewardAddress, empAbi]"
     );
     assert(
-      empValue.length === 2,
-      "Each EMP whitelisted is expected to be an array in the form [empAddress, rewardAddress]"
+      empValue.length >= 3,
+      "Each EMP whitelisted is expected to be an array in the form [empAddress, rewardAddress, empAbi]"
     );
     return empValue;
   }
@@ -210,6 +215,7 @@ module.exports = ({ queries, empAbi, coingecko, synthPrices, firstEmpDate }) => 
           if (err.message.includes("history does not go back far enough") || err.message.includes("history is empty")) {
             result.push([empAddress, "0"]);
           } else {
+            console.error("error with emp", empAddress);
             throw err;
           }
         }
@@ -299,7 +305,7 @@ module.exports = ({ queries, empAbi, coingecko, synthPrices, firstEmpDate }) => 
         collateralTokens,
         async address => await getCoingeckoPriceHistory(address, "usd", startTime, endTime)
       ),
-      Promise.map(
+      Promise.mapSeries(
         empWhitelist,
         // each empwhitelist contains, [empaddress,deployeraddres], so we only care about empaddress
         async ([empAddress], i) =>
@@ -314,7 +320,7 @@ module.exports = ({ queries, empAbi, coingecko, synthPrices, firstEmpDate }) => 
       getBlocks(startTime, endTime),
       // Each empwhitelist contains, [empaddress,deployeraddres], balance histories expects array of just emp addresses
       getAllBalanceHistories(
-        empWhitelist.map(([empaddress]) => empaddress),
+        empWhitelist.map(([empaddress, , empAbi]) => [empaddress, empAbi]),
         firstEmpDate,
         endTime
       )